% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_task.R
\docType{data}
\name{heart.model}
\alias{heart.model}
\title{Heart Classification Drive ML Model.}
\format{
An object of class \code{autoMLmodel} of length 6.
}
\usage{
heart.model
}
\value{
heart data driveML sample model output
}
\description{
Contains the task (`heart.model`).
}
\examples{
## Sample model object
modelobj <- heart.model
}
\references{
See \url{https://www.kaggle.com/cdabakoglu/heart-disease-classifications-machine-learning}
}
\keyword{mlmodel}
