% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printAutoDataPrep.R
\name{printautoDataprep}
\alias{printautoDataprep}
\title{Print Method for the autoDataprep Class}
\usage{
printautoDataprep(x)
}
\arguments{
\item{x}{[Object | Required] an object of class autoDataprep}
}
\value{
Print summary autoDataprep fucntion results on consloe
}
\description{
Print the result of autoDataprep object
}
\examples{
#Auto data prep
traindata <- autoDataprep(heart, target = "target_var", missimpute = "default",
dummyvar = TRUE, aucv = 0.02, corr = 0.98, outlier_flag = TRUE,
interaction_var = TRUE, frequent_var = TRUE)

# Print auto data prep object
printautoDataprep(traindata)
}
