% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printAutoMAR.R
\name{printautoMAR}
\alias{printautoMAR}
\title{Print Method for the autoMAR Class
Print the result of autoMAR object}
\usage{
printautoMAR(x)
}
\arguments{
\item{x}{[Object | Required] an object of class autoMAR}
}
\value{
Print summary of autoMAR output in console
}
\description{
Print Method for the autoMAR Class
Print the result of autoMAR object
}
\examples{
## Missing at random features
marobj <- autoMAR (heart, aucv = 0.9, strataname = NULL, stratasize = NULL, mar_method = "glm")

## print summary in console
printautoMAR(marobj)
}
