% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoFeature.R
\name{generateFeature}
\alias{generateFeature}
\title{Automated column transformer}
\usage{
generateFeature(data, varlist, type = "Frequent", method = NULL)
}
\arguments{
\item{data}{dataframe or data.table}

\item{varlist}{variable list to generate the additional features}

\item{type}{variable transformation type "'Dummy','Outlier','Frequent', 'Interaction', 'svd', 'cluster', 'cluster distance', 'text', 'WOE', 'PCA'"}

\item{method}{input for variabe transforamtion. For type = 'Frequent' then type should be 'Frequency' or 'Percent'. Other type method list is provided in details}
}
\value{
generated transformed features
}
\description{
This function automatically scans through each variable and generate features based on the type listed in the detail
}
