% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printAutoMAR.R
\name{printautoMAR}
\alias{printautoMAR}
\title{Print Method for the autoMAR Class
Print the result of autoMAR object}
\usage{
printautoMAR(x)
}
\arguments{
\item{x}{[Object | Required] an object of class autoMAR}
}
\description{
Print Method for the autoMAR Class
Print the result of autoMAR object
}
